- added saveTag test 0003266: error when removing all rights in shared tag edit dialog http://www.tine20.org/bugtracker/view.php?id=3266
git-svn-id: file:///www/metaways/svn.tine20.org/repo/trunk@16916 8d11cdf3-9d2f-0410-bbb5-e098f187f06b

+++ tests/tine20/Admin/JsonTest.php
@@ -117,7 +117,12 @@ class Admin_JsonTest extends PHPUnit_Framework_TestCase
     {
         // remove accounts for group member tests
         try {
-            Admin_Controller_User::getInstance()->delete($this->objects['user']->accountId);
+            if (array_key_exists('user', $this->objects)) {
+                Admin_Controller_User::getInstance()->delete($this->objects['user']->accountId);
+            }
+            if (array_key_exists('tag', $this->objects)) {
+                Admin_Controller_Tags::getInstance()->delete($this->objects['tag']['id']);
+            }
         } catch (Exception $e) {
             // do nothing
         }
@@ -568,5 +573,40 @@ class Admin_JsonTest extends PHPUnit_Framework_TestCase
         $allRights = $this->_backend->getAllRoleRights();
         
         $this->assertGreaterThan(0, $allRights);
+    }
+    
+    /**
+     * try to save tag
+     */
+    public function testSaveTag()
+    {
+        $tagData = $this->_getTagData();
+        $this->objects['tag'] = $this->_backend->saveTag($tagData);
+        $this->assertEquals($tagData['name'], $this->objects['tag']['name']);
+        //print_r($this->objects['tag']);
+    }
+    
+    /**
+     * get tag data
+     * 
+     * @return array
+     */
+    protected function _getTagData()
+    {
+        return array(
+            'rights' => array(
+                array(
+                    'account_id' => 0,
+                    'account_type' => 'anyone',
+                    'account_name' => 'Anyone',
+                    'view_right' => true,
+                    'use_right' => true
+                )
+            ),
+            'contexts' => array('any'),
+            'name' => 'supertag',
+            'description' => 'xxxx',
+            'color' => '#003300'
+        );
     }
 }
